Change call to worldwide_api to parse a hash
Affects all smartanswers

gds-api-adapters returns a hash of the results from the worldwide_api,
not an objects with callable methods.

Converting the response from worldwide_api to an OpenStruct allows us to
treat the keys in the hash as methods. So we can still use the delegate
to get the details. However, OpenStruct does not allow you to elegantly
do recursive conversion of a hash to an OpenStruct, so as a quick fix,
I have removed the delegate for the slug and created a method to extract
it from details.