Refactor call to imminence_api
Following the changes from gds-api-adapters gem where it now returns a hash
of the results from the imminence_api, I have made the changes in this commit
to properly handle the new format.

First by using the to_hash call from gds-api-adapters to convert the response to a ruby
hash data structure. An openstruct is then to wrap the hash converting its keys to properties.

This makes it easier to resolve cases where the results key may not be returned, gracefully
returning nil.