Preparatory: Add Brittish or Irish option
This commit shows the question and outcomes after including the
british-or-irish option to the regression test configuration files.

This addition triggers the creation of british-or-irish regression artefacts.

This gives assurance that this setup will form the basis for further future
changes in later commits.

See these documents [1] and [1] for more details

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned commit/changes have had no
(interfering) effect to other smart answers.

[1] https://docs.google.com/document/d/1PtZxv9Fy-SOUtRSDvUq7_jbqLYUBn8Pp0nCf57U2UIE/edit
[2] https://docs.google.com/drawings/d/1x2547adffe616JMZ6qnI2ST5H2f-OnnUwSZ5isn5dzM/edit