Preparatory: Add Netherlands
This commit update the question and responses manually with the
addition of Netherlands to list of options for country_of_birth? question

This is in preparation for the content changes to be made in later commit
where a user gives birth in the Netherlands and lives in North Korea.

This commit also shows the affected/changes to regression test artefacts
as a result of the changes made to add new countries.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.